Revert "MLT-846.2"
