--- .zeplin/components.json
@@ -1,5 +1,0 @@
-{
-  "projects": ["5cf658d06af3e11bb49146d9"],
-  "styleguides": [],
-  "components": []
-}
