+++ LICENSE.MD
@@ -630,7 +630,7 @@ state the exclusion of warranty; and each file should have at least
 the "copyright" line and a pointer to where the full notice is found.
 
     <The first interface to the new IATI Datastore interfacing all Grand Bargain signatories>
-    Copyright (C) 2019 Zimmerman
+    Copyright (C) 2019  Zimmerman & Zimmerman
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU Affero General Public License as published
