+++ src/app/modules/signatory-data/mock.ts
@@ -34,7 +34,7 @@ export const iatigbsignatoriesCallValues = {
                 type: 'query',
                 q:
                   '(humanitarian_scope_vocabulary:"2-1" AND humanitarian_scope_code:*) 
-                  OR (sector_code:* AND sector_vocabulary:10) OR (transaction_sector_code:* 
+                  OR (sector:* AND sector_vocabulary:10) OR (transaction_sector_code:* 
                   AND transaction_sector_vocabulary:10) 
                   OR (humanitarian_scope_vocabulary:"1-2" AND humanitarian_scope_code:*)',
               },
