+++ src/app/modules/signatory-data/submodules/activityList/const.tsx
@@ -19,7 +19,13 @@ export const activitiesQuery = (
     q: `reporting_org_ref:${repOrgRef} AND title_narrative:${searchTerm} AND 
       (${actStatusFilters}) AND ${selCountries}${
       extraFilter ? ` ${extraFilter} AND ` : ''
-    } (humanitarian:1 OR transaction_humanitarian:1 OR (-(-sector_vocabulary:1 OR sector_vocabulary:*) AND (sector_code:[70000 TO 79999] OR sector_code:[93010 TO 93018])) OR (-(-transaction_sector_vocabulary:1 OR transaction_sector_vocabulary:*) AND (transaction_sector_code:[70000 TO 79999] OR transaction_sector_code:[93010 TO 93018])))`,
+    }
+      (humanitarian:1 OR transaction_humanitarian:1 OR 
+      (-(-sector_vocabulary:1 OR sector_vocabulary:*) AND 
+      (sector_code:[70000 TO 79999] OR sector_code:[93010 TO 93018])) OR 
+      (-(transaction_sector_vocabulary:1 OR transaction_sector_vocabulary:*) AND 
+      (transaction_sector_code:[70000 TO 79999] OR
+       transaction_sector_code:[93010 TO 93018])))`,
     fl: `iati_identifier,activity_status_code,title,recipient_country_code,transaction_recipient_country_code,activity_date_start_actual,activity_date_end_actual,result_type,activity_date_start_planned,activity_date_end_planned`,
   };
 };
