+++ src/app/modules/signatory-data/submodules/overview/const.ts
@@ -7,7 +7,7 @@ export const humCallValues = {
   humElData_2: {
     type: 'query',
     q:
-      '(sector_vocabulary:10 AND sector_code:*) OR (transaction_sector_vocabulary:10 AND transaction_sector_code:*)',
+      '(sector_vocabulary:10 AND sector:[* TO *]) OR (transaction_sector_vocabulary:10 AND transaction_sector_code:*)',
   },
   humElData_3: {
     type: 'query',
@@ -178,7 +178,7 @@ export const barJsonFacet = (years, queryDateField) => {
   years.forEach(year => {
     result[year] = {
       type: 'query',
-      q: `activity_date_start_actual_f:[${year}-01-01T00:00:00Z TO ${year}-12-31T23:59:59Z] OR activity_date_start_planned_f:[${year}-01-01T00:00:00Z TO ${year}-12-31T23:59:59Z]`,
+      q: `activity_date_start_actual:[${year}-01-01T00:00:00Z TO ${year}-12-31T24:00:00Z]' OR activity_date_start_planned:[${year}-01-01T00:00:00Z TO ${year}-12-31T24:00:00Z]'`,
       facet: {
         hum_count: {
           type: 'query',
