+++ src/app/modules/signatory-data/submodules/overview/index.tsx
@@ -63,15 +63,6 @@ export function OverviewPage(props) {
     state => state.sigdataoverviewdataerrors.data
   );
   const tooltipsData = useStoreState(globalState => globalState.tooltips.data);
-  const loading = useStoreState(
-    state =>
-      state.sigdataactivityyears.loading ||
-      state.sigdataactivitiesbyyear.loading ||
-      state.sigdataactivitystatus.loading ||
-      state.sigdataoverviewhum.loading ||
-      state.sigdataoverviewhum4donut.loading ||
-      state.sigdataoverviewcurrency.loading
-  );
   /* create the API call instances */
   const sigdataactivitiesbyyearCall = useStoreActions(
     actions => actions.sigdataactivitiesbyyear.fetch
@@ -254,7 +245,6 @@ export function OverviewPage(props) {
       humOtherClassOfInterestData={humOtherClassOfInterestData}
       humActwGBClassificationsData={humActwGBClassificationsData}
       description={description}
-      loading={loading}
     />
   );
 }
