+++ src/app/modules/signatory-data/submodules/overview/model.ts
@@ -30,5 +30,4 @@ export interface OverviewLayoutModel {
   activityTimelineData: YearBarChartObjectModel[];
   humanitarianElementsData: HumanitarianDonutModel[];
   description: string;
-  loading: boolean;
 }
